% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPairedEnd.R
\name{readPairedEnd}
\alias{readPairedEnd}
\title{pull in (filter and assemble as an object) some or all of the "useful" reads
in a paired-end sequencing run}
\usage{
readPairedEnd(
  bam,
  genome = c("hg19", "hg38", "mm10", "mm9"),
  bamParams = NULL,
  which = NULL,
  ...
)
}
\arguments{
\item{bam}{character string, the BAM file to parse}

\item{genome}{optional character string, the genome assembly to target}

\item{bamParams}{optional any parameters to pass through to ScanBamParam}

\item{which}{optional a GRanges object with specific regions to extract}
}
\value{
a GenomicAlignmentPairs object
}
\description{
pull in (filter and assemble as an object) some or all of the "useful" reads
in a paired-end sequencing run
}
