% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBAM.R
\name{readBAM}
\alias{readBAM}
\title{pull in (filter and assemble as an object) some or all of the "useful" reads
in a paired-end sequencing run}
\usage{
readBAM(
  bam,
  genome = c("hg19", "hg38", "mm10", "mm9"),
  is.single = FALSE,
  bamParams = NULL,
  which = NULL,
  style = "UCSC",
  ...
)
}
\arguments{
\item{bam}{character string, the BAM file to parse}

\item{genome}{optional character string, the genome assembly to target}

\item{is.single}{whethr the input BAM is single-end or not}

\item{bamParams}{optional any parameters to pass through to ScanBamParam}

\item{which}{optional a GRanges object with specific regions to extract}

\item{style}{what style the assembly annotations are in (e.g. 'chr1' == UCSC or '1' == NCBI)}
}
\value{
a GenomicAlignmentPairs object
}
\description{
pull in (filter and assemble as an object) some or all of the "useful" reads
in a paired-end sequencing run
}
